+++ core/outbound/clients/resolver/tcp_resolver.go
@@ -40,7 +40,7 @@ func (r *TCPResolver) Init() error {
 		r.poolConn, err = r.createConnectionPool(
 			func() (net.Conn, error) { return r.CreateBaseConn() })
 		if err != nil {
-			log.Debugf("Set %s pool's IdleTimeout to %d, InitialCapacity to %d, MaxCapacity to %d", r.dnsUpstream.Name, r.dnsUpstream.TCPPoolConfig.IdleTimeout, r.dnsUpstream.TCPPoolConfig.InitialCapacity, r.dnsUpstream.TCPPoolConfig.MaxCapacity)
+			log.Warnf("Failed to create TCP connection pool for %s: %s", r.dnsUpstream.Name, err)
 		}
 	} else {
 		return nil